city_layers_app/requirements.txt
streamlit
folium
streamlit-folium
requests
python-dotenv
fastapi
uvicorn
